+++ TimescaleManager/DTO/TimescaleValueDTO.cs
@@ -3,7 +3,7 @@
     public class TimescaleValueDTO
     {
         public DateTime Date { get; set; }
-        public double ExecutionTime { get; set; }
-        public double Value { get; set; }
+        public float ExecutionTime { get; set; }
+        public float Value { get; set; }
     }
 }
